+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - fixed improves method setIptables from wirelessmode default mode 
 - fixed clear dependencies not used from requirements files
 - fixed include message: the module not found or failed to import.
+- fixed typo name proxys to proxies
 
 
 ## [released]